+++ src/components/Modal/AlertModal/StyledAlertModal.ts
@@ -1,42 +1,28 @@
 import styled from 'styled-components';
 
-const AlertModalWrap = styled.div`
-  .modalContent {
-    background: var(--background-color);
-    border-radius: 1rem;
-    width: 25.2rem;
-    font-size: var(--text-l);
-    overflow: hidden;
-    position: relative;
-  }
-
-  .modalList {
-    button {
-      width: 50%;
-      text-align: center;
-      padding: 1.2rem;
-      font-size: var(--text-m);
-      transition: all 0.2s ease-in-out;
-    }
-
-    button:first-child {
-      border-right: 1px solid var(--gray-200);
-    }
-
-    button:last-child {
-      color: var(--point-dark-400);
-    }
-
-    button:hover {
-      background-color: var(--point-color);
-    }
-  }
-`;
-
-const Header = styled.header`
-  padding: 2.2rem;
-  border-bottom: 1px solid var(--gray-200);
+const StyledAlertModal = styled.dialog`
+  max-width: min(300px, calc(100% - var(--margin-mobile) * 2));
   text-align: center;
+  border-radius: 1rem;
+  background: var(--background-color);
+
+  h3 {
+    padding: 22px 30px;
+    font-size: var(--text-l);
+    word-break: keep-all;
+  }
+
+  button {
+    width: 100%;
+    padding: 1.2rem;
+    font-size: var(--text-m);
+    color: var(--point-dark-400);
+    border-top: 1px solid var(--gray-200);
+  }
+
+  button:focus {
+    outline-color: var(--primary-color);
+  }
 `;
 
-export { AlertModalWrap, Header };
+export default StyledAlertModal;
